Fix some UDP doc problems
(1) UDP is a connectionless protocol. To refer to a connection creates an incorrect impression and invalid expectations in the mind of the programmer. Fixed some occurrences of this. 
(2) Some functions were documented as being mandatory to call after parsePacket(). Wrong. What was meant is that the function returns info determined by parsePacket().