Properly link between the hardware datasheet pages
Currently when clicking the Core button from the Photon datasheet page crashes and clicking the Photon button from the Core datasheet page shows the old Photon datasheet page.

This PR adds new fields to the front matter to allow alternate links for the Photon/Core button at the top of the page (`core_link` and `photon_link`). This can be extended later to add similar links for the Electron.

Any page that exists only in the Core folder should include an appropriate `photon_link`, and vise-versa.

Also fixed here is the Edit link for the `ios.md` page.