Document how to expose a C++ method to the cloud
The firmware supports exposing a method on an object with
`Spark.function` by accepting a `std::function` but figuring out how to
do it requires C++ template ninja skills.

This PR adds an example on using `std::bind` to expose the method of a
C++ object through `Spark.function`.

The example was tested to work on a Photon.