Update firmware.md
Added "pinMode(D2, INPUT);" to attachinterrupt code example. If the D2 pin is not declared as an input pin, nothing happens with the attachinterrupt() (at least on a core). Since Arduino code does not require this declaration, not having this in the example is a recipe for disappointment and wasted time for those that come with this assumption that the photon/core will be just like an arduino.