Update examples.md
Your configuration file was backwards, the code says to place your hand in front of the beam, but then we measure the on values, next we are told to remove our hand, yet we measure the off values. I switched the on and off values so we measure off when the hand is in front and on when the hand is removed.