Example improvement
This example wasn't copy&paste compatible, and the `Udp.begin(8888);` part was missing. Without this, no data will be sent. It is true that this line was mentioned in a example before, but not here. This caused confusion for me, and other people on the internets. 

I hope this helps.