Give fix-it tip for Cloud not connected
Background: Our code uses TCPClient, and in its current stage of development, it often goes into the GREEN stage, and we need to put it into safe mode to flash the new code. We were very perplexed when it first happened, and nearly ordered a new Photon as we didn't understand why it wasn't cloud-connected. We figured it out after some forum hunting though.

So my proposal is to put a tip in the device modes doc, based off the most frequent tip in the forum.
Y'all likely know more about other causes of Cloud not being connected, so you may have better wording.

Thanks!