Update iOS link
The link to the iPhone app is wrong. It should have been the "Particle - Build IoT projects - WiFi or Cellular" app instead of "Particle Core", which seems to be the old app for Core, otherwise the connection to the Photon device would always fail. It took me 1 hour to figure this out.