Updated the first paragraph of "Particle.connect()" function
First paragraph of Particle.connect() for electron device was saying:
"Particle.connect() connects the device to the Cloud. This will automatically activate the Wi-Fi module and attempt to connect to a Wi-Fi network if the device is not already connected to a network."

I replaced the mention of WiFi to cellular connection.