Update LED blink code description
The "Flashing cyan" blink sequence happens when you are connected to a Wifi network (i.e. have an IP), but have not yet connected to the internet. I'm not sure if this is 100% correct or if there are other scenarios where this blink sequence will happen, but that is what I have observed with the Photon.

This might be redundant with the Flash red twice error code, but I'm not sure how long it takes for that error code to appear. Perhaps adding a description of a timeout will clear that up?