Keep titles in lowercase in the reference section
Several pages in the reference section have headers that are commands (CLI) or function names (JS client). Transforming them to uppercase makes it harder for users to find the correct capitalization.

Also switch the section id to the lowercase filename because it was generating stuff like `id="Command Line Interface (CLI)-content"` otherwise.